requirements.txt
langchain
langchain_community
huggingface_hub
sentence-transformers
chromadb
python-dotenv